requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP client
httpx>=0.28.1,<1

# AI Services
assemblyai>=0.52.0,<1
openai>=1.40.0,<2
google-genai==1.33.0

# Environment
python-dotenv==1.0.1

# Utilities
pydantic==2.5.3
certifi>=2024.0.0
